ref(AlertRuleStatus): Split <AlertRuleStatus> into a component in its own file
